# Binomial: Faster sampling for n * p < 10