Fixes #734.