@dhardy 2019-02-20
Performance looks decent (< 1ms for all n tested):
```
bench-binomial$ cargo run  
    Finished dev [unoptimized + debuginfo] target(s) in 0.35s
     Running `target/debug/bench-binomial`
1.00e0  Rand: 1000000 in 0.000s Statrs: 1000000 in 0.388s       GSL: 1000000 in 0.002s
1.00e-1 Rand: 99845 in 0.000s   Statrs: 100194 in 0.388s        GSL: 99915 in 0.000s
1.00e-2 Rand: 9889 in 0.000s    Statrs: 10081 in 0.404s GSL: 10129 in 0.000s
1.00e-3 Rand: 1006 in 0.000s    Statrs: 950 in 0.402s   GSL: 1024 in 0.000s
1.00e-4 Rand: 109 in 0.000s     Statrs: 96 in 0.423s    GSL: 108 in 0.000s
1.00e-5 Rand: 14 in 0.000s      Statrs: 10 in 0.401s    GSL: 9 in 0.000s
1.00e-6 Rand: 1 in 0.000s       Statrs: 1 in 0.416s     GSL: 0 in 0.000s
1.00e-7 Rand: 0 in 0.000s       Statrs: 1 in 0.393s     GSL: 1 in 0.000s
1.00e-8 Rand: 0 in 0.000s       Statrs: 0 in 0.409s     GSL: 0 in 0.000s
1.00e-9 Rand: 0 in 0.000s       Statrs: 0 in 0.386s     GSL: 0 in 0.000s
...
1.00e-31        Rand: 0 in 0.000s       Statrs: 0 in 0.385s     GSL: 0 in 0.000s
1.00e-32        Rand: 0 in 0.000s       Statrs: 0 in 0.384s     GSL: 0 in 0.000s
1.00e-33        Rand: 0 in 0.000s       Statrs: 0 in 0.384s     GSL: 0 in 0.000s
1.00e-34        Rand: 0 in 0.000s       Statrs: 0 in 0.386s     GSL: 0 in 0.000s
1.00e-35        Rand: 0 in 0.000s       Statrs: 0 in 0.384s     GSL: 0 in 0.000s
1.00e-36        Rand: 0 in 0.000s       Statrs: 0 in 0.387s     GSL: 0 in 0.000s
...
1.00e-296       Rand: 0 in 0.000s       Statrs: 0 in 0.400s     GSL: 0 in 0.000s
1.00e-297       Rand: 0 in 0.000s       Statrs: 0 in 0.401s     GSL: 0 in 0.000s
1.00e-298       Rand: 0 in 0.000s       Statrs: 0 in 0.414s     GSL: 0 in 0.000s
1.00e-299       Rand: 0 in 0.000s       Statrs: 0 in 0.416s     GSL: 0 in 0.000s
1.00e-300       Rand: 0 in 0.000s       Statrs: 0 in 0.403s     GSL: 0 in 0.000s
1.00e-301       Rand: 0 in 0.000s       Statrs: 0 in 0.400s     GSL: 0 in 0.000s
1.00e-302       Rand: 0 in 0.000s       Statrs: 0 in 0.408s     GSL: 0 in 0.000s
1.00e-303       Rand: 0 in 0.000s       Statrs: 0 in 0.416s     GSL: 0 in 0.000s
1.00e-304       Rand: 0 in 0.000s       Statrs: 0 in 0.406s     GSL: 0 in 0.000s
1.00e-305       Rand: 0 in 0.000s       Statrs: 0 in 0.406s     GSL: 0 in 0.000s
1.00e-306       Rand: 0 in 0.000s       Statrs: 0 in 0.398s     GSL: 0 in 0.000s
1.00e-307       Rand: 0 in 0.000s       Statrs: 0 in 0.401s     GSL: 0 in 0.000s
1.00e-308       Rand: 0 in 0.000s       Statrs: 0 in 0.401s     GSL: 0 in 0.000s
1.00e-309       Rand: 0 in 0.000s       Statrs: 0 in 0.400s     GSL: 0 in 0.000s
```