@vks 2019-02-21
> As noted, it would be nice to look into other approaches also, but that can be a future PR.

I "implemented" the recommended algorithm for large expected values in the paper, but it does not quite work yet. It is a bit of the pain that the algorithm is specified with gotos.

I agree that this should be another PR however.