@vks 2019-02-25
By using integer exponentiation (using the `num` implementation modified to work with `u64` exponents), I was able to get a significant speedup:

Before:
```
test distr_binomial       ... bench:      88,365 ns/iter (+/- 2,215) = 90 MB/s
test distr_binomial_small ... bench:      33,775 ns/iter (+/- 5,494) = 236 MB/s
test misc_binomial_1      ... bench:          13 ns/iter (+/- 2)
test misc_binomial_10     ... bench:          72 ns/iter (+/- 1)
test misc_binomial_100    ... bench:          71 ns/iter (+/- 11)
test misc_binomial_1000   ... bench:         624 ns/iter (+/- 19)
test misc_binomial_1e12   ... bench:         681 ns/iter (+/- 18)
```

After:
```
test distr_binomial       ... bench:      44,354 ns/iter (+/- 3,518) = 180 MB/s
test distr_binomial_small ... bench:      22,736 ns/iter (+/- 514) = 351 MB/s
test misc_binomial_1      ... bench:          10 ns/iter (+/- 0)
test misc_binomial_10     ... bench:          23 ns/iter (+/- 0)
test misc_binomial_100    ... bench:          27 ns/iter (+/- 4)
test misc_binomial_1000   ... bench:         621 ns/iter (+/- 15)
test misc_binomial_1e12   ... bench:         686 ns/iter (+/- 20)
```
Is this worth it to ship our own `powi` implementation? It's only 14 lines of code.