@dhardy 2019-02-27
Hope I didn't merge this too soon; this test appears to have hung: https://travis-ci.org/rust-random/rand/jobs/498834909